# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Goodfood 3.0 Users Microservice",
    "version": "1.0.0",
    "title": "Goodfood 3.0 Users Service API"
  },
...
  "paths": {
...
    "/subscription": {
      "post": {
        "tags": ["subscription"],
        "summary": "Create subscription",
        "operationId": "createSubscription",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subscription created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          },
          "400": {
            "description": "Invalid form input or required fields missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Referenced resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error contacting external service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/user/{id}/subscription": {
      "get": {
        "tags": ["subscription"],
        "summary": "Get a user's subscriptions",
        "operationId": "getUserSubscriptions",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The ID of the user who's subscriptions we're looking for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User subscriptions found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Subscription"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/user/{id}/subscription/last-cancelled": {
      "get": {
        "tags": ["subscription"],
        "summary": "Get the user's last cancelled subscription",
        "operationId": "getLastCancelledSubscription",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "type",
            "description": "Optional filter for type of subscription cancelled. Pass \"SCHEDULED\" for Mealkit subsciptions and \"MEMBERSHIP\" for WOW subscriptions. Passing nothing finds the last cancelled subscription from all types of subscriptions.",
            "schema": {
              "$ref": "#/components/schemas/BaseSubscriptionInput/properties/subscription_type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user's last cancelled subscription. Returns null if user has not cancelled a subscription before.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/LastCancelledSubscription"
                    },
                    {
                      "type": "object",
                      "nullable": true
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid subscription type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/subscription/{id}/weekday": {
      "patch": {
        "tags": ["subscription"],
        "summary": "Update a SCHEDULED subscription's delivery day",
        "operationId": "updateSubscriptionDeliveryDay",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The subscription ID to be updated",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "delivery_day": {
                    "$ref": "#/components/schemas/BaseSubscriptionInput/properties/delivery_day"
                  }
                },
                "required": ["delivery_day"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription weekday updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          },
          "400": {
            "description": "Invalid form input or required fields missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Subscription not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Valid request submitted, but request could not be processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error contacting external service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/subscription/{id}/plan": {
      "patch": {
        "tags": ["subscription"],
        "summary": "Update a SCHEDULED subscription's plan",
        "operationId": "updateSubscriptionPlan",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The subscription ID to be updated",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plan_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "old_plan_id": {
                    "type": "number",
                    "description": "GF2 legacy plan ID"
                  }
                },
                "required": ["plan_id", "old_plan_id"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription plan updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          },
          "400": {
            "description": "Invalid form input or required fields missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Subscription not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Valid request submitted, but request could not be processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error contacting external service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/subscription/cancellation_reason": {
      "get": {
        "summary": "Cancellation reasons",
        "tags": ["subscription"],
        "operationId": "getCancellationReasons",
        "description": "Returns a list of all possible cancellation reasons",
        "parameters": [
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "is_user_visible",
            "description": "If true, only shows user visible reasons. If false, only shows non user-visible reasons."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/CancellationReasonValue"
                      },
                      {
                        "$ref": "#/components/schemas/CancellationReasonEntity"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a reason for cancellation",
        "operationId": "createCancellationReason",
        "tags": ["subscription"],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CancellationReasonValue"
                    },
                    {
                      "$ref": "#/components/schemas/CancellationReasonEntity"
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancellationReasonValue"
              }
            }
          },
          "description": ""
        },
        "description": "Create a reason for cancellation"
      },
      "parameters": []
    },
    "/subscription/cancellation_reason/{id}": {
      "parameters": [
        {
          "schema": {
            "type": "number"
          },
          "name": "id",
          "in": "path",
          "description": "ID of the reason",
          "required": true
        }
      ],
      "put": {
        "summary": "Update a reason for cancellation",
        "tags": ["subscription"],
        "operationId": "updateCancellationReason",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CancellationReasonEntity"
                    },
                    {
                      "$ref": "#/components/schemas/CancellationReasonValue"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "description": "Update a reason for cancellation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancellationReasonValue"
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["subscription"],
        "summary": "Delete a reason for cancellation",
        "operationId": "deleteCancellationReason",
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "description": "Deletes a reason for cancellation"
      }
    },
    "/subscription/{id}/cancellation": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of the subscription"
        }
      ],
      "get": {
        "summary": "Get a list of reasons why this subscription was cancelled",
        "tags": ["subscription"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/CancellationSelectionEntity"
                      },
                      {
                        "$ref": "#/components/schemas/CancellationSelectionValue"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "operationId": "getCancellationSelections",
        "description": "Returns the reasons for this subscription's cancellation"
      },
      "post": {
        "summary": "",
        "tags": ["subscription"],
        "operationId": "PostSubscriptionCancellation",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/CancellationSelectionEntity"
                      },
                      {
                        "$ref": "#/components/schemas/CancellationSelectionValue"
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "example-1": {
                    "value": {
                      "error": "Subscription not found"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error contacting external service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CancellationSelectionValue"
                }
              }
            }
          },
          "description": "An array of selections over cancellation_reason"
        },
        "description": "Sets the subscription into cancellation state with an array of reason selections in the body."
      }
    },
    "/subscription/{id}/membership": {
      "get": {
        "tags": ["subscription"],
        "summary": "Get subscription membership",
        "description": "",
        "operationId": "getSubscriptionMembership",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "The ID of the subscription to fetch membership for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Membership"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "price": {
                          "$ref": "#/components/schemas/MembershipPrice"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Subscription not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Wrong subscription type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "MembershipPrice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "membership_id": {
            "type": "string",
            "format": "uuidv4"
          },
          "province_code": {
            "type": "string",
            "enum": ["QC", "ON", "NL", "PE", "NS", "NB", "MB", "SK", "AB", "BC", "YT", "NT", "NU"]
          },
          "price": {
            "type": "number"
          },
          "tax_code": {
            "type": "string"
          }
        },
        "required": ["id", "membership_id", "province_code", "price", "tax_code"]
      },
...
      "SubscriptionInput": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseSubscriptionInput"
          },
          {
            "type": "object",
            "properties": {
              "ip_address": {
                "oneOf": [
                  {
                    "type": "string",
                    "format": "ipv4"
                  },
                  {
                    "type": "string",
                    "format": "ipv6"
                  }
                ]
              },
              "coupon_code": {
                "type": "string",
                "example": "MEALKIT45"
              }
            },
            "required": ["ip_address"]
          }
        ]
      },
      "Subscription": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuidv4"
          },
          "user_id": {
            "type": "string",
            "format": "uuidv4"
          },
          "card_id": {
            "type": "string",
            "format": "uuidv4"
          },
          "address_id": {
            "type": "string",
            "format": "uuidv4"
          },
          "agreement_id": {
            "type": "integer"
          },
          "state": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "CANCELLATION",
              "TRIAL",
              "UNPAID",
              "CANCELLED",
              "PAUSED",
              "SUSPENDED"
            ]
          },
          "subscription_type": {
            "type": "string",
            "enum": ["SCHEDULED", "MEMBERSHIP"]
          },
          "product_id": {
            "type": "string",
            "format": "uuidv4"
          },
          "frequency_type": {
            "type": "string",
            "enum": ["DAY", "MONTH", "YEAR"]
          },
          "frequency_value": {
            "type": "integer"
          },
          "delivery_day": {
            "type": "string",
            "enum": ["MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY", "SUNDAY"]
          },
          "next_cycle": {
            "type": "string",
            "format": "date"
          },
          "send_notification": {
            "type": "boolean"
          },
          "started_at": {
            "type": "string",
            "format": "date"
          },
          "product": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Membership"
              },
              {
                "$ref": "#/components/schemas/Plan"
              }
            ]
          }
        },
        "required": [
          "id",
          "user_id",
          "card_id",
          "address_id",
          "state",
          "subscription_type",
          "product_id",
          "frequency_type",
          "frequency_value",
          "next_cycle",
          "send_notification",
          "started_at"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "error description"
          },
          "meta": {
            "description": "Structured representation of the error",
            "example": {
              "errors": [["The fsa field is required."]]
            }
          },
          "status_code": {
            "type": "number"
          }
        }
      },
      "CancellationReasonValue": {
        "description": "",
        "type": "object",
        "title": "CancellationReasonProperties",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1
          },
          "entry_type": {
            "type": "string",
            "minLength": 1,
            "enum": ["STATIC", "EDITABLE"],
            "default": "STATIC"
          },
          "priority": {
            "type": "integer",
            "default": null,
            "nullable": true
          },
          "is_user_visible": {
            "type": "boolean"
          }
        },
        "required": ["code", "is_user_visible"]
      },
      "CancellationReasonEntity": {
        "title": "CancellationReasonEntityProperties",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "CancellationSelectionValue": {
        "title": "CancellationSelectionValue",
        "type": "object",
        "x-examples": {},
        "properties": {
          "reason_id": {
            "type": "integer",
            "example": 2
          },
          "edit_value": {
            "type": "string",
            "nullable": true,
            "example": "User stopped eating food"
          },
          "category": {
            "type": "string",
            "nullable": true,
            "example": "Contentful category"
          },
          "source": {
            "type": "string",
            "nullable": true,
            "enum": ["ADA", "CLIENT", "SYSTEM", "ZENDESK"],
            "example": "CLIENT"
          }
        },
        "description": "A user's selection for value"
      },
      "CancellationSelectionEntity": {
        "title": "CancellationSelectionEntity",
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "subscription_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "LastCancelledSubscription": {
        "type": "object",
        "description": "The last cancelled subscription by the user",
        "properties": {
          "create_payload": {
            "description": "Payload to pass to createSubscription. old_plan_id will be added by BFF.",
            "$ref": "#/components/schemas/BaseSubscriptionInput"
          },
          "product": {
            "description": "Details on the membership or meal plan",
            "oneOf": [
              {
                "$ref": "#/components/schemas/Membership"
              },
              {
                "$ref": "#/components/schemas/Plan"
              }
            ]
          }
        }
      },
...
    },
...
  }
}